refactor: move getRandomExpiration to setup.ts as requested
- Move getRandomExpiration function from test/utils/utils.ts to test/integration/setup.ts
- Update imports in postOrder.spec.ts and privateListing.spec.ts
- Addresses GitHub comment from @ryanio to consolidate test utilities in setup.ts
