+++ test/utils/utils.ts
@@ -26,17 +26,3 @@ export const expectValidOrder = (order: OrderV2) => {
     expect(field in order).to.be.true;
   }
 };
-
-export const getRandomExpiration = (): number => {
-  const now = Math.floor(Date.now() / 1000);
-  const fifteenMinutes = 15 * 60;
-  const oneHour = 60 * 60;
-  const range = oneHour - fifteenMinutes + 1;
-
-  const crypto = require("crypto");
-  const randomBuffer = crypto.randomBytes(4);
-  const randomValue = randomBuffer.readUInt32BE(0);
-  const randomSeconds = (randomValue % range) + fifteenMinutes;
-
-  return now + randomSeconds;
-};
